requirements.txt
# Video downloading and scraping
yt-dlp>=2024.1.1

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# XG file parsing
zlib-wrapper>=0.1.3

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.25.0

# Video processing
opencv-python>=4.8.0
decord>=0.6.0
av>=10.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
